+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed Magunet retaining it's charge when switched or dropped while charging
 - Fixed Magunet charge particles getting stuck in the air when switched or dropped while charging
 - Fixed softlock caused by getting hit after dropping Warrior's Gi and starting a new floor
+- Fixed Ticonderogun being able to hit invulnerable enemies
 
 ## 1.11.2 (2024-04-07)